feat: Add import_ministeriali action to LibriController
This commit adds a new action, `import_ministeriali`, to the `LibriController` class. The action handles the import of ministeriali books from a SQL file. If the import is successful, the user is redirected to the libri index page with a success notice. Otherwise, they are redirected with an error alert.